Added static_cast to "double to int" warning
git-svn-id: https://svnsrv.sdu.dk/svn/RobWork/RobWork/trunk@391 494c09c7-a67d-49d4-9938-151fb1965be8

+++ src/rwlibs/pathplanners/prm/PartialIndexTable.hpp
@@ -51,7 +51,7 @@ private:
         size_t index = 0;
         for (typename std::vector<Dimension>::iterator it = _dimensions.begin(); it != _dimensions.end(); ++it) {
             Dimension dim = *it;
-            size_t i1 = (q(dim.index)-dim.qoffset)/dim.stepsize;
+            size_t i1 = static_cast<size_t>((q(dim.index)-dim.qoffset)/dim.stepsize);
             index += i1*dim.inc;
         }
         return index;
